Remove InlineFlowDisplayListBuilding
This lets us remove a few call indirections where a function just calls
another function defined in another module.
